refactor: extract JWT validation into dedicated file
Move JWT-specific validation helpers out of `string.go` to improve organization and maintainability. Update spelling and workspace checksum metadata.
